requirements.txt
pyOpenSSL==17.3.0
ecdsa==0.13